Revert "gralloc: Add support for CRC calculation"
Qcom introduced these interfaces in LA.QSSI.11.0, but not in 12.0
and 13.0, maybe because they are not needed for future chipsets.
As a result, only lahaina needs those interfaces.

Revert this commit such that it can be built with LA.QSSI.13.0, which
are tags for commonsys display libs.

Thanks Jake Weinstein for the information.

This reverts commit 10d758df9d3ff2fab47d0f5dd1272f6cf6f7fbdd.

Change-Id: I11542facc60376f0179fb70a71585a70c37bef82